+++ src/lib.rs
@@ -1,12 +1,13 @@
 use ordered_float::OrderedFloat;
 
+mod candle;
 mod candlestick_chart;
 mod candlestick_chart_state;
 mod symbols;
 mod x_axis;
 mod y_axis;
 
-pub use candlestick_chart::Candle;
+pub use candle::Candle;
 pub use candlestick_chart::CandleStickChart;
 pub use candlestick_chart_state::CandleStickChartState;
